+++ test/support/postgres/domain.ex
@@ -10,5 +10,6 @@ defmodule AshWorkflowTest.Postgres.Domain do
     resource AshWorkflowTest.Postgres.ScreeningTransition
     resource AshWorkflowTest.Postgres.AuthorizedScreeningWorkflow
     resource AshWorkflowTest.Postgres.ScreeningWithTimeoutWorkflow
+    resource AshWorkflowTest.Postgres.RepeatUntilWorkflow
   end
 end